Change "grabbed" to "contents" to reflect the direction of relationship
Words like "grabbed" and "held" can imply the opposite of what is
happening. For example, someone can hold your hand, and your hand can
be grabbed. Referring to the contents of your hand is a one-way
relationship, even if it sounds a bit generic. It also opens the way for
the contents to have a reference to the container as well, naturally
describing the relationship.