Add ability to compile fennel in the lib and src directories
The code that exists needs to be split out and I think using the Rust
convention of putting most of the code in lib so that it's easily
testable with most of the application setup done from files in src is a
good way to go. Of course, LOVR dictates that the main file must be in
the root directory, but it can immediately turn around and run a file in
src to maintain overall project organization.