Move the wrap utility to lib/math.fnl
This is just the beginning - tests are soon to follow.

As a side note, you can't have a module that has the same name as a
built-in one because you won't be able to load it. The package will
already exist in the table of loaded packages so it won't even open the
file. Putting all code in a non-root directory ensures that these
namespace conflicts can't happen because everything will be prefixed
with lib or src.