Draw block connections more clearly
This draws the outgoing lines from the right side of the source into the
left side of the destination. This makes it much more clear what the
order of the links are, and no longer hides cycles between two blocks.
(Ultimately that shouldn't happen, but the representation is bound to
change at least once in the future and I think that will prevent that
problem entirely, so no real need to fix it right now.)