Treat user input as a layer module
This marks the first time that code written in VR could alter the VR
environment other than calling the host code directly! It's also wrapped
in a logging breaker which means it shouldn't be able to crash the host
environment.

I think some of the next things that are needed are better block
grabbing (I can't tell you how often I grabbed the _wrong block_ because
the block grabbing algorithm is a hack), some kind of toggle layer with
a safety input sequence to toggle back, and some hooks for the user code
to be able to alter the programming environment so that things like
the text input mode can be changed from within VR, possibly to things
that were written in VR.