Allow breakers to accept nil as a layer
Generally this is useless in normal code since you'd never intentionally
create a breaker with a nil layer, but if the user forgets to return a
layer compatible table, then this could happen accidentally. Since the
breaker doesn't protect its own code, this means that the environment
hard crashes. This prevents that.