Upgrade fennel to 0.9.2 and use APE lua
This replaces the fennel executable with a lua interpreter and the
fennel script because the 0.9.2 executable gets flagged by Windows
Defender as a trojan. Bundling a lua interpreter means that upgrading
fennel in the future should be easier, since it should just overwrite
the existing fennel script instead of changing out executables.

The bundled lua interpreter is an actually portable executable [1] build
of lua, which means it should just work on more than just Windows,
making the build more portable too.

[1] https://justine.lol/ape.html